AMB-1909 extract python timer
