+++ lambda_code/src/models/fhir_immunization.py
@@ -1,31 +1,11 @@
 """Immunization FHIR R4B validator"""
-import logging
-import time
-
-from functools import wraps
 
 from typing import Literal
 from fhir.resources.R4B.immunization import Immunization
 from models.fhir_immunization_pre_validators import FHIRImmunizationPreValidators
 from models.fhir_immunization_post_validators import FHIRImmunizationPostValidators
 from models.utils.generic_utils import get_generic_questionnaire_response_value
-
-logging.basicConfig()
-logger = logging.getLogger("my-logger")
-logger.setLevel(logging.DEBUG)
-
-def timed(func):
-    """This decorator prints the execution time for the decorated function."""
-
-    @wraps(func)
-    def wrapper(*args, **kwargs):
-        start = time.time()
-        result = func(*args, **kwargs)
-        end = time.time()
-        logger.debug("{} ran in {}s".format(func.__name__, round(end - start, 2)))
-        return result
-
-    return wrapper
+from timer import timed
 
 class ImmunizationValidator:
     """
